the stable version v14b
This code is for theoretically studing transient electron transports in heterojunctions of gold metals and silicon quantum dots by using the time-dependent non-equilibrium green function (TDNEGF) method. The studied quantum dots are Si-SiO2 core-shell nanoparticles that are doped with phosphorus impurities for low-voltage applications. In order to quickly compute devices having more than a thousand atoms, we demonstrate an analytical algorithm based on muffin-tin approximation to solve the Poisson equation for transient variations of electron densities, and apply the wide-band limit approximation for contacts of electrodes. Numerical results are described and discussed herein.

Please refer to the article:
http://arxiv.org/abs/1810.02577